Previously, we treated the collection as array which is highly
inefficient. This was a regression from version 0.5.

before:

```
[1 / 4] Count all of 1000 root entities...
  7.136ms (±14.37%) per iteration
  37s elapsed (18% for setup) for 4198 iterations in 16 cycles

[2 / 4] Count all of 1000000 root entities...
  5688.454ms (±63.62%) per iteration
  49s elapsed (31% for setup) for 6 iterations in 6 cycles

[3 / 4] Count about half of 1000 root entities...
  14.501ms (±5.17%) per iteration
  31s elapsed (2% for setup) for 2069 iterations in 13 cycles

[4 / 4] Count about half of 1000000 root entities...
  2383.355ms (±27.08%) per iteration
  41s elapsed (30% for setup) for 12 iterations in 7 cycles
```


now:

```
[1 / 4] Count all of 1000 root entities...
5.451ms (±1.82%) per iteration
21s elapsed (3% for setup) for 3710 iterations in 9 cycles

[2 / 4] Count all of 1000000 root entities...
7.426ms (±6.03%) per iteration
40s elapsed (25% for setup) for 4046 iterations in 10 cycles

[3 / 4] Count about half of 1000 root entities...
7.213ms (±1.99%) per iteration
16s elapsed (2% for setup) for 2172 iterations in 7 cycles

[4 / 4] Count about half of 1000000 root entities...
968.006ms (±16.43%) per iteration
43s elapsed (30% for setup) for 31 iterations in 14 cycles
``